// src/types.d.ts
export type Phone = {
  name: string | null
  price: number | null
  screenSize: number | null
  batteryCapacity: number | null
  brand: string | null
  storage: number | null
  ram: number | null
  cameraMpx: number | null
  cpu: string | null
  chargingPortType: string | null
  fingerprint: boolean | null
  os: string | null 
}